Config file builder  completely moved to separate file.
